requirements.txt
streamlit>=1.30.0
google-analytics-data>=0.18.0
google-auth>=2.25.0
google-api-python-client>=2.110.0
pandas>=2.1.0
plotly>=5.18.0
python-dotenv>=1.0.0
